teamonboard page ui updated

+++ src/views/TeamOnboard.vue
@@ -1,50 +1,94 @@
 <template>
-  <v-container>
-    <v-row justify="center" align="center">
-      <v-col cols="12" md="4" v-if="isVisible && (!error || error.length < 1)">
-        <v-text-field
-          class="ma-0 pa-0"
-          label="Email"
-          v-model="userData.email"
-          disabled
-          outlined
-        ></v-text-field>
-        <v-text-field
-          class="ma-0 pa-0"
-          label="Password"
-          type="password"
-          v-model="userData.password"
-          outlined
-        ></v-text-field>
-        <v-btn
-          @click="onBoardUser"
-          class="primary"
-          style="height: 38px; width: 150px"
-          :loading="loading"
-          depressed
-          >Onboard</v-btn
+  <v-main>
+    <v-container class="mt-md-15">
+      <v-row justify="center" align="center">
+        <v-col
+          cols="12"
+          md="4"
+          lg="3"
+          sm="6"
+          class="text-center mt-md-5"
+          v-if="isVisible && (!error || error.length < 1)"
         >
-      </v-col>
-      <v-col sm="12" v-else-if="!successDialog && error.length > 1">
-        {{ error }}
-      </v-col>
-    </v-row>
+          <p class="google-font" style="font-size:2.5em">Team Onboard</p>
+          <v-divider class="mb-md-10"></v-divider>
+          <br />
+          <v-text-field
+            class="ma-0 pa-0"
+            label="Email"
+            v-model="userData.email"
+            disabled
+            outlined
+          ></v-text-field>
+          <v-text-field
+            class="ma-0 pa-0"
+            label="Password"
+            v-model="userData.password"
+            :append-icon="show1 ? 'mdi-eye' : 'mdi-eye-off'"
+            :type="show1 ? 'text' : 'password'"
+            @click:append="show1 = !show1"
+            placeholder="Your Password"
+            outlined
+          ></v-text-field>
+          <v-text-field
+            class="ma-0 pa-0"
+            label="Confirm Password"
+            :append-icon="show1 ? 'mdi-eye' : 'mdi-eye-off'"
+            :type="show2 ? 'text' : 'password'"
+            @click:append="show2 = !show2"
+            v-model="cpassword"
+            placeholder="Confirm Password"
+            outlined
+          ></v-text-field>
+          <v-btn
+            @click="onBoardUser"
+            class="primary"
+            style="height: 38px; width: 150px"
+            :loading="loading"
+            depressed
+            >Onboard</v-btn
+          >
+        </v-col>
+        <v-col
+          sm="4"
+          class="google-font text-center"
+          v-else-if="!successDialog && error.length > 1"
+        >
+          <p class="google-font" style="font-size:2.5em">Team Onboard</p>
+          <v-divider></v-divider>
 
-    <v-dialog v-model="successDialog" width="500">
-      <v-card>
-        <v-card-title class="headline grey lighten-2"> Success </v-card-title>
+          <p class="mt-md-10 mt-4" style="font-size:1.2em">{{ error }}</p>
+          <v-btn class="mt-3" to="/" color="primary" depressed
+            >Back to Login</v-btn
+          >
+        </v-col>
+      </v-row>
 
-        <v-card-text> Your account is created </v-card-text>
+      <v-dialog v-model="successDialog" width="500">
+        <v-card>
+          <v-card-title
+            class="google-font grey lighten-2"
+            style="font-size:1.3em"
+          >
+            Success
+          </v-card-title>
 
-        <v-divider></v-divider>
+          <v-card-text>
+            <p class="google-font mt-4 mb-0" style="font-size:1.2em">
+              Your account is created
+            </p>
+          </v-card-text>
 
-        <v-card-actions>
-          <v-spacer></v-spacer>
-          <v-btn color="primary" text @click="dialogOkay"> Okay </v-btn>
-        </v-card-actions>
-      </v-card>
-    </v-dialog>
-  </v-container>
+          <v-divider></v-divider>
+
+          <v-card-actions>
+            <v-spacer></v-spacer>
+            <v-btn color="primary" text @click="dialogOkay"> Okay </v-btn>
+          </v-card-actions>
+        </v-card>
+      </v-dialog>
+    </v-container>
+  </v-main>
 </template>
 
 <script>
@@ -53,10 +97,13 @@ import firebase from "@/config/firebase";
 export default {
   name: "teamonboard",
   data: () => ({
+    show1: false,
+    show2: false,
     loading: false,
     successDialog: false,
     isVisible: false,
     error: "",
+    cpassword: "",
     finalUser: {},
     userData: {
       password: "",
@@ -79,29 +126,34 @@ export default {
   methods: {
     async onBoardUser() {
       this.loading = true;
-      try {
-        const firebaseUser = await firebase.auth.createUserWithEmailAndPassword(
-          this.finalUser.email,
-          this.userData.password
-        );
-        await firebase.auth.signOut();
-        const deleteData = await teamOnboardService.deleteTempUse(
-          this.$route.params.id
-        );
-        if (!deleteData.success) throw { message: "Something went wrong" };
-        const data = { uid: firebaseUser.user.uid, ...this.finalUser };
-        const userAdd = await teamOnboardService.addOnBoardUserToUserManagement(
-          data
-        );
-        const updateTeamFields = await teamOnboardService.updateTeamWithUID(
-          data.uid,
-          data.id
-        );
-        if (!updateTeamFields.success) throw updateTeamFields;
-        this.loading = true;
-        this.successDialog = true;
-      } catch (e) {
-        console.log(e);
+      if (this.cpassword == this.userData.password) {
+        try {
+          const firebaseUser = await firebase.auth.createUserWithEmailAndPassword(
+            this.finalUser.email,
+            this.userData.password
+          );
+          await firebase.auth.signOut();
+          const deleteData = await teamOnboardService.deleteTempUse(
+            this.$route.params.id
+          );
+          if (!deleteData.success) throw { message: "Something went wrong" };
+          const data = { uid: firebaseUser.user.uid, ...this.finalUser };
+          const userAdd = await teamOnboardService.addOnBoardUserToUserManagement(
+            data
+          );
+          const updateTeamFields = await teamOnboardService.updateTeamWithUID(
+            data.uid,
+            data.id
+          );
+          if (!updateTeamFields.success) throw updateTeamFields;
+          this.loading = true;
+          this.successDialog = true;
+        } catch (e) {
+          console.log(e);
+          this.loading = false;
+        }
+      }else{
+        alert('Password Not Match')
         this.loading = false;
       }
     },
@@ -114,5 +166,4 @@ export default {
 };
 </script>
 
-<style>
-</style>
+<style></style>
